+++ net/tcp/tcp_timer.c
@@ -364,7 +364,7 @@ void tcp_stop_timer(FAR struct tcp_conn_s *conn)
  *
  ****************************************************************************/
 
-void tcp_set_zero_probe(FAR struct tcp_conn_s *conn, uint16_t flags)
+void tcp_set_zero_probe(FAR struct tcp_conn_s *conn, uint32_t flags)
 {
   if ((conn->tcpstateflags & TCP_ESTABLISHED) &&
       ((flags & TCP_NEWDATA) == 0) && conn->tx_unacked <= 0 &&
