+++ net/usrsock/usrsock_poll.c
@@ -47,8 +47,8 @@
  * Private Functions
  ****************************************************************************/
 
-static uint16_t poll_event(FAR struct net_driver_s *dev,
-                           FAR void *pvpriv, uint16_t flags)
+static uint32_t poll_event(FAR struct net_driver_s *dev,
+                           FAR void *pvpriv, uint32_t flags)
 {
   FAR struct usrsock_poll_s *info = pvpriv;
   FAR struct usrsock_conn_s *conn = info->conn;
